flaskheroku/requirements.txt
firebase-admin
pyzbar
Flask
flask-cors
requests
Pillow